requirements.txt
# Core Data Science Libraries
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Deep Learning & NLP
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
accelerate>=0.20.0

# API & Web Server
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0

# Configuration & Environment
python-dotenv>=1.0.0
pyyaml>=6.0

# Testing
pytest>=7.4.0
httpx>=0.24.0

# Utilities
tqdm>=4.65.0
joblib>=1.3.0

# Optional: Jupyter for notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
